+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Rename ApplicationForm's `@action` and `@method` to `@default_action`, `@default_method`
 * Exclude namespaces from model_name when inheriting form objects
+* Fix Rails compatibility bug in `ApplicationForm#as_json`
 
 ## 0.1.0